+++ clients/admin-ui/src/features/datastore-connections/system_portal_config/forms/ConnectorParameters.tsx
@@ -53,9 +53,8 @@ const createSaasConnector = async (
   systemFidesKey: string,
   createSaasConnectorFunc: any
 ) => {
-  const connectionConfig: CreateSaasConnectionConfigRequest = {
+  const connectionConfig: Omit<CreateSaasConnectionConfigRequest, "name"> = {
     description: values.description,
-    name: values.name,
     instance_key: formatKey(values.instance_key as string),
     saas_connector_type: connectionOption.identifier,
     secrets: {},
@@ -91,16 +90,16 @@ export const patchConnectionConfig = async (
       ? formatKey(values.instance_key as string)
       : connectionConfig?.key;
 
-  const params1: Omit<ConnectionConfigurationResponse, "created_at"> = {
-    access: AccessLevel.WRITE,
-    connection_type: (connectionOption.type === SystemType.SAAS
-      ? connectionOption.type
-      : connectionOption.identifier) as ConnectionType,
-    description: values.description,
-    disabled: false,
-    key,
-    name: values.name,
-  };
+  const params1: Omit<ConnectionConfigurationResponse, "created_at" | "name"> =
+    {
+      access: AccessLevel.WRITE,
+      connection_type: (connectionOption.type === SystemType.SAAS
+        ? connectionOption.type
+        : connectionOption.identifier) as ConnectionType,
+      description: values.description,
+      disabled: false,
+      key,
+    };
   const payload = await patchFunc({
     systemFidesKey,
     connectionConfigs: [params1],