+++ tests/ops/api/v1/endpoints/test_encryption_endpoints.py
@@ -5,18 +5,18 @@ from unittest.mock import Mock
 import pytest
 from starlette.testclient import TestClient
 
-from fides.api.api.v1.urn_registry import (
-    DECRYPT_AES,
-    ENCRYPT_AES,
-    ENCRYPTION_KEY,
-    V1_URL_PREFIX,
-)
 from fides.api.cryptography.cryptographic_util import b64_str_to_bytes, bytes_to_b64_str
 from fides.api.util.encryption.aes_gcm_encryption_scheme import (
     decrypt,
     encrypt_verify_secret_length,
 )
 from fides.common.api.scope_registry import ENCRYPTION_EXEC, STORAGE_CREATE_OR_UPDATE
+from fides.common.api.v1.urn_registry import (
+    DECRYPT_AES,
+    ENCRYPT_AES,
+    ENCRYPTION_KEY,
+    V1_URL_PREFIX,
+)
 from fides.config import CONFIG